mailfilter.py: add new script use standlib to handle maildir and message
This script is to replace mailgrep.py. Use the standard lib of python3,
mailbox, to access the messages in the Maildir format mailbox. And, use
the standard lib email.message to handle the message.
The process is esaier and robuster than the code in mailgrep.py. Python
should be more esay.

Signed-off-by: Navy Cheng <navych@126.com>